🚀 Move code to modules/base.py
